backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyMuPDF
openai
anthropic
google-generativeai
groq
mistralai
python-dotenv
